+++ spark/v3.1/spark/src/main/scala/com/netease/arctic/spark/sql/execution/ExtendedIcebergStrategy.scala
@@ -18,26 +18,21 @@
 
 package com.netease.arctic.spark.sql.execution
 
-import com.netease.arctic.spark.sql.catalyst.plans.{AppendArcticData, OverwriteArcticData, OverwriteArcticDataByExpression, ReplaceArcticData}
-import com.netease.arctic.spark.table.ArcticSparkTable
 import com.netease.arctic.spark.{ArcticSparkCatalog, ArcticSparkSessionCatalog}
 import org.apache.iceberg.spark.{Spark3Util, SparkCatalog, SparkSessionCatalog}
 import org.apache.spark.sql.catalyst.analysis.NamedRelation
-import org.apache.spark.sql.catalyst.expressions.{And, Expression, NamedExpression, PredicateHelper}
+import org.apache.spark.sql.catalyst.expressions.{And, Expression, NamedExpression}
 import org.apache.spark.sql.catalyst.planning.PhysicalOperation
-import org.apache.spark.sql.catalyst.plans.CreateArcticTableAsSelect
 import org.apache.spark.sql.catalyst.plans.logical._
-import org.apache.spark.sql.catalyst.utils.TranslateUtils
 import org.apache.spark.sql.connector.catalog.{Identifier, TableCatalog}
 import org.apache.spark.sql.connector.iceberg.read.SupportsFileFilter
 import org.apache.spark.sql.execution.datasources.v2._
 import org.apache.spark.sql.execution.{FilterExec, LeafExecNode, ProjectExec, SparkPlan}
-import org.apache.spark.sql.util.CaseInsensitiveStringMap
 import org.apache.spark.sql.{SparkSession, Strategy}
 
-import scala.collection.JavaConverters.{mapAsJavaMapConverter, seqAsJavaList}
+import scala.collection.JavaConverters.seqAsJavaList
 
-case class ExtendedIcebergStrategy(spark: SparkSession) extends Strategy with PredicateHelper{
+case class ExtendedIcebergStrategy(spark: SparkSession) extends Strategy {
   override def apply(plan: LogicalPlan): Seq[SparkPlan] = plan match {
     case DynamicFileFilter(scanPlan, fileFilterPlan, filterable) =>
       DynamicFileFilterExec(planLater(scanPlan), planLater(fileFilterPlan), filterable) :: Nil
@@ -58,42 +53,6 @@ case class ExtendedIcebergStrategy(spark: SparkSession) extends Strategy with Pr
 
     case ReplaceData(relation, batchWrite, query) =>
       ReplaceDataExec(batchWrite, refreshCache(relation), planLater(query)) :: Nil
-
-    case ReplaceArcticData(table: DataSourceV2Relation, query, options) =>
-      table.table match {
-        case t: ArcticSparkTable =>
-            AppendDataExec(t, new CaseInsensitiveStringMap(options.asJava), planLater(query), refreshCache(table)) :: Nil
-      }
-
-    case AppendArcticData(table: DataSourceV2Relation, query, validateQuery, options) =>
-      table.table match {
-        case t: ArcticSparkTable =>
-          AppendInsertDataExec(t, new CaseInsensitiveStringMap(options.asJava), planLater(query),
-            planLater(validateQuery), refreshCache(table)) :: Nil
-      }
-
-    case OverwriteArcticData(table: DataSourceV2Relation, query, validateQuery, options) =>
-      table.table match {
-        case t: ArcticSparkTable =>
-          OverwriteArcticDataExec(t, new CaseInsensitiveStringMap(options.asJava), planLater(query),
-            planLater(validateQuery), refreshCache(table)) :: Nil
-      }
-
-    case OverwriteArcticDataByExpression(table: DataSourceV2Relation, deleteExpr, query, validateQuery, options) =>
-      table.table match {
-        case t: ArcticSparkTable =>
-          val filters = splitConjunctivePredicates(deleteExpr).map {
-            filter => TranslateUtils.translateFilter(deleteExpr).getOrElse(
-              throw new UnsupportedOperationException("Cannot translate expression to source filter"))
-          }.toArray
-          OverwriteArcticByExpressionExec(t, filters, new CaseInsensitiveStringMap(options.asJava), planLater(query),
-            planLater(validateQuery), refreshCache(table)) :: Nil
-      }
-
-    case CreateArcticTableAsSelect(catalog, ident, parts, query, validateQuery, props, options, ifNotExists) =>
-      val writeOptions = new CaseInsensitiveStringMap(options.asJava)
-      CreateArcticTableAsSelectExec(catalog, ident, parts, query, planLater(query), planLater(validateQuery),
-        props, writeOptions, ifNotExists) :: Nil
 
     case MergeInto(mergeIntoParams, output, child) =>
       MergeIntoExec(mergeIntoParams, output, planLater(child)) :: Nil
